Fix command line server parameter
dropping backward compatibility with older formats.
