+++ pkg/tsdbctl/tsdbctl_test.go
@@ -37,8 +37,8 @@ type testTsdbctlSuite struct {
 }
 
 func (suite *testTsdbctlSuite) TestPopulateConfigWithTenant() {
-	rc := RootCommandeer{v3ioPath: "Vel@Odar:p455w0rd@localhost:80123/123"}
-	cfg := &config.V3ioConfig{Username: "Vel@Odar", Password: "p455w0rd", TablePath: "/x/y/z"}
+	rc := RootCommandeer{v3ioUrl: "localhost:80123"}
+	cfg := &config.V3ioConfig{Username: "Vel@Odar", Password: "p455w0rd", Container: "123", TablePath: "/x/y/z"}
 
 	err := rc.populateConfig(cfg)
 	suite.Require().Nil(err)
@@ -52,11 +52,11 @@ func (suite *testTsdbctlSuite) TestPopulateConfigWithTenant() {
 
 	expectedRc := RootCommandeer{
 		v3iocfg:  cfg,
-		v3ioPath: "localhost:80123/123",
+		v3ioUrl:  "localhost:80123",
 		Reporter: metricReporter,
 	}
 	expectedCfg := &config.V3ioConfig{
-		WebApiEndpoint: "localhost:80123",
+		WebApiEndpoint: "http://localhost:80123",
 		Container:      "123",
 		TablePath:      "/x/y/z",
 		Username:       "Vel@Odar",
@@ -79,12 +79,12 @@ func (suite *testTsdbctlSuite) TestContainerConfig() {
 	suite.Require().NoError(err)
 	defer os.Setenv("V3IO_ACCESS_KEY", oldAccessKey)
 
-	rc := RootCommandeer{v3ioPath: "Vel@Odar:p455w0rd@localhost:80123/123", container: "test", accessKey: "acce55-key"}
+	rc := RootCommandeer{v3ioUrl: "localhost:80123/123", container: "test", accessKey: "acce55-key"}
 	cfg := &config.V3ioConfig{Username: "Vel@Odar", Password: "p455w0rd", TablePath: "/x/y/z"}
 
 	err = rc.populateConfig(cfg)
 	expectedCfg := &config.V3ioConfig{
-		WebApiEndpoint: "localhost:80123",
+		WebApiEndpoint: "http://localhost:80123",
 		Container:      "test",
 		TablePath:      "/x/y/z",
 		Username:       "Vel@Odar",
@@ -114,7 +114,7 @@ func (suite *testTsdbctlSuite) TestConfigFromEnvVarsAndPassword() {
 
 	err = rc.populateConfig(cfg)
 	expectedCfg := &config.V3ioConfig{
-		WebApiEndpoint: "host-from-env:123",
+		WebApiEndpoint: "http://host-from-env:123",
 		Container:      "test",
 		TablePath:      "/x/y/z",
 		Username:       "Vel@Odar",
@@ -143,7 +143,7 @@ func (suite *testTsdbctlSuite) TestConfigFromEnvVars() {
 
 	err = rc.populateConfig(cfg)
 	expectedCfg := &config.V3ioConfig{
-		WebApiEndpoint: "host-from-env:123",
+		WebApiEndpoint: "http://host-from-env:123",
 		Container:      "test",
 		TablePath:      "/x/y/z",
 		AccessKey:      "key-from-env",
